+++ codec/collections.go
@@ -398,10 +398,11 @@ func protoCol(f protoreflect.FieldDescriptor) schema.Field {
 			col.Kind = schema.StringKind
 		case protoreflect.BytesKind:
 			col.Kind = schema.BytesKind
+			col.Nullable = true
 		case protoreflect.EnumKind:
 			// TODO: support enums
 			col.Kind = schema.EnumKind
-			// use the full name to avoid collissions
+			// use the full name to avoid collisions
 			col.ReferencedType = string(f.Enum().FullName())
 			col.ReferencedType = strings.ReplaceAll(col.ReferencedType, ".", "_")
 		case protoreflect.MessageKind: