+++ runtime/v2/module.go
@@ -18,6 +18,7 @@ import (
 	"cosmossdk.io/core/branch"
 	"cosmossdk.io/core/comet"
 	"cosmossdk.io/core/event"
+	"cosmossdk.io/core/gas"
 	"cosmossdk.io/core/header"
 	"cosmossdk.io/core/registry"
 	"cosmossdk.io/core/router"
@@ -147,7 +148,6 @@ func ProvideAppBuilder[T transaction.Tx](
 type AppInputs struct {
 	depinject.In
 
-	StoreConfig        *root.Config
 	Config             *runtimev2.Module
 	AppBuilder         *AppBuilder[transaction.Tx]
 	ModuleManager      *MM[transaction.Tx]
@@ -214,6 +214,7 @@ func ProvideEnvironment(
 	kvService store.KVStoreService,
 	memKvService store.MemoryStoreService,
 	headerService header.Service,
+	gasService gas.Service,
 	eventService event.Service,
 	branchService branch.Service,
 	routerBuilder RouterServiceBuilder,
@@ -222,7 +223,7 @@ func ProvideEnvironment(
 		Logger:             logger,
 		BranchService:      branchService,
 		EventService:       eventService,
-		GasService:         stf.NewGasMeterService(),
+		GasService:         gasService,
 		HeaderService:      headerService,
 		QueryRouterService: routerBuilder.BuildQueryRouter(),
 		MsgRouterService:   routerBuilder.BuildMsgRouter([]byte(key.Name())),
@@ -296,6 +297,7 @@ func DefaultServiceBindings() depinject.Config {
 		eventService                = services.NewGenesisEventService(stf.NewEventService())
 		storeBuilder                = root.NewBuilder()
 		branchService               = stf.BranchService{}
+		gasService                  = stf.NewGasMeterService()
 	)
 	return depinject.Supply(
 		kvServiceFactory,
@@ -305,5 +307,6 @@ func DefaultServiceBindings() depinject.Config {
 		eventService,
 		storeBuilder,
 		branchService,
+		gasService,
 	)
 }