+++ simapp/v2/simdv2/cmd/commands.go
@@ -1,6 +1,7 @@
 package cmd
 
 import (
+	"context"
 	"io"
 
 	"github.com/spf13/cobra"
@@ -23,11 +24,13 @@ import (
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/client/config"
 	"github.com/cosmos/cosmos-sdk/client/debug"
+	"github.com/cosmos/cosmos-sdk/client/grpc/cmtservice"
+	nodeservice "github.com/cosmos/cosmos-sdk/client/grpc/node"
 	"github.com/cosmos/cosmos-sdk/client/keys"
 	"github.com/cosmos/cosmos-sdk/client/rpc"
 	sdktelemetry "github.com/cosmos/cosmos-sdk/telemetry"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/types/module"
+	txtypes "github.com/cosmos/cosmos-sdk/types/tx"
 	"github.com/cosmos/cosmos-sdk/version"
 	authcmd "github.com/cosmos/cosmos-sdk/x/auth/client/cli"
 	"github.com/cosmos/cosmos-sdk/x/genutil"
@@ -111,8 +114,12 @@ func InitRootCmd[T transaction.Tx](
 			simApp.Name(),
 			simApp.Store(),
 			simApp.App.AppManager,
-			simApp.AppCodec(),
-			&client.DefaultTxDecoder[T]{TxConfig: deps.TxConfig},
+			cometbft.AppCodecs[T]{
+				AppCodec:              simApp.AppCodec(),
+				TxCodec:               &client.DefaultTxDecoder[T]{TxConfig: deps.TxConfig},
+				LegacyAmino:           deps.ClientContext.LegacyAmino,
+				ConsensusAddressCodec: deps.ClientContext.ConsensusAddressCodec,
+			},
 			simApp.App.QueryHandlers(),
 			simApp.App.SchemaDecoderResolver(),
 			initCometOptions[T](),
@@ -149,17 +156,12 @@ func InitRootCmd[T transaction.Tx](
 		logger,
 		deps.GlobalConfig,
 		simApp.InterfaceRegistry(),
+		simApp.App.AppManager,
 	)
 	if err != nil {
 		return nil, err
 	}
-
-	for _, mod := range deps.ModuleManager.Modules() {
-		if gmod, ok := mod.(module.HasGRPCGateway); ok {
-			// TODO(@julienrbrt) https://github.com/cosmos/cosmos-sdk/pull/22701#pullrequestreview-2470651390
-			gmod.RegisterGRPCGatewayRoutes(deps.ClientContext, grpcgatewayServer.GRPCGatewayRouter)
-		}
-	}
+	registerGRPCGatewayRoutes[T](deps, grpcgatewayServer)
 
 	// wire server commands
 	return serverv2.AddCommands[T](
@@ -263,4 +265,17 @@ func RootCommandPersistentPreRun(clientCtx client.Context) func(*cobra.Command, 
 
 		return nil
 	}
+}
+
+// registerGRPCGatewayRoutes registers the gRPC gateway routes for all modules and other components
+// TODO(@julienrbrt): Eventually, this should removed and directly done within the grpcgateway.Server
+// ref: https://github.com/cosmos/cosmos-sdk/pull/22701#pullrequestreview-2470651390
+func registerGRPCGatewayRoutes[T transaction.Tx](
+	deps CommandDependencies[T],
+	server *grpcgateway.Server[T],
+) {
+	// those are the extra services that the CometBFT server implements (server/v2/cometbft/grpc.go)
+	cmtservice.RegisterGRPCGatewayRoutes(deps.ClientContext, server.GRPCGatewayRouter)
+	_ = nodeservice.RegisterServiceHandlerClient(context.Background(), server.GRPCGatewayRouter, nodeservice.NewServiceClient(deps.ClientContext))
+	_ = txtypes.RegisterServiceHandlerClient(context.Background(), server.GRPCGatewayRouter, txtypes.NewServiceClient(deps.ClientContext))
 }