+++ store/v2/root/store.go
@@ -8,8 +8,6 @@ import (
 	"sync"
 	"time"
 
-	"golang.org/x/sync/errgroup"
-
 	corelog "cosmossdk.io/core/log"
 	corestore "cosmossdk.io/core/store"
 	"cosmossdk.io/store/v2"
@@ -116,6 +114,12 @@ func (s *Store) getVersionedReader(version uint64) (store.VersionedReader, error
 	return nil, fmt.Errorf("version %d does not exist", version)
 }
 
+// StateAt returns a read-only view of the state at a given version.
+func (s *Store) StateAt(v uint64) (corestore.ReaderMap, error) {
+	vReader, err := s.getVersionedReader(v)
+	return NewReaderMap(v, vReader), err
+}
+
 func (s *Store) StateLatest() (uint64, corestore.ReaderMap, error) {
 	v, err := s.GetLatestVersion()
 	if err != nil {
@@ -129,12 +133,6 @@ func (s *Store) StateLatest() (uint64, corestore.ReaderMap, error) {
 	return v, NewReaderMap(v, vReader), nil
 }
 
-// StateAt returns a read-only view of the state at a given version.
-func (s *Store) StateAt(v uint64) (corestore.ReaderMap, error) {
-	vReader, err := s.getVersionedReader(v)
-	return NewReaderMap(v, vReader), err
-}
-
 func (s *Store) GetStateCommitment() store.Committer {
 	return s.stateCommitment
 }
@@ -144,7 +142,7 @@ func (s *Store) GetStateCommitment() store.Committer {
 // latest version set, which is based off of the SC view.
 func (s *Store) LastCommitID() (proof.CommitID, error) {
 	if s.lastCommitInfo != nil {
-		return s.lastCommitInfo.CommitID(), nil
+		return *s.lastCommitInfo.CommitID(), nil
 	}
 
 	latestVersion, err := s.stateCommitment.GetLatestVersion()
@@ -171,8 +169,7 @@ func (s *Store) GetLatestVersion() (uint64, error) {
 
 func (s *Store) Query(storeKey []byte, version uint64, key []byte, prove bool) (store.QueryResult, error) {
 	if s.telemetry != nil {
-		now := time.Now()
-		defer s.telemetry.MeasureSince(now, "root_store", "query")
+		defer s.telemetry.MeasureSince(time.Now(), "root_store", "query")
 	}
 
 	val, err := s.stateCommitment.Get(storeKey, version, key)
@@ -198,8 +195,7 @@ func (s *Store) Query(storeKey []byte, version uint64, key []byte, prove bool) (
 
 func (s *Store) LoadLatestVersion() error {
 	if s.telemetry != nil {
-		now := time.Now()
-		defer s.telemetry.MeasureSince(now, "root_store", "load_latest_version")
+		defer s.telemetry.MeasureSince(time.Now(), "root_store", "load_latest_version")
 	}
 
 	lv, err := s.GetLatestVersion()
@@ -212,8 +208,7 @@ func (s *Store) LoadLatestVersion() error {
 
 func (s *Store) LoadVersion(version uint64) error {
 	if s.telemetry != nil {
-		now := time.Now()
-		defer s.telemetry.MeasureSince(now, "root_store", "load_version")
+		defer s.telemetry.MeasureSince(time.Now(), "root_store", "load_version")
 	}
 
 	return s.loadVersion(version, nil, false)
@@ -221,8 +216,7 @@ func (s *Store) LoadVersion(version uint64) error {
 
 func (s *Store) LoadVersionForOverwriting(version uint64) error {
 	if s.telemetry != nil {
-		now := time.Now()
-		defer s.telemetry.MeasureSince(now, "root_store", "load_version_for_overwriting")
+		defer s.telemetry.MeasureSince(time.Now(), "root_store", "load_version_for_overwriting")
 	}
 
 	return s.loadVersion(version, nil, true)
@@ -235,7 +229,6 @@ func (s *Store) LoadVersionAndUpgrade(version uint64, upgrades *corestore.StoreU
 	if upgrades == nil {
 		return errors.New("upgrades cannot be nil")
 	}
-
 	if s.telemetry != nil {
 		defer s.telemetry.MeasureSince(time.Now(), "root_store", "load_version_and_upgrade")
 	}
@@ -293,7 +286,9 @@ func (s *Store) loadVersion(v uint64, upgrades *corestore.StoreUpgrades, overrid
 func (s *Store) Commit(cs *corestore.Changeset) ([]byte, error) {
 	if s.telemetry != nil {
 		now := time.Now()
-		defer s.telemetry.MeasureSince(now, "root_store", "commit")
+		defer func() {
+			s.telemetry.MeasureSince(now, "root_store", "commit")
+		}()
 	}
 
 	if err := s.handleMigration(cs); err != nil {
@@ -305,25 +300,17 @@ func (s *Store) Commit(cs *corestore.Changeset) ([]byte, error) {
 	// background pruning process (iavl v1 for example) which must be paused during the commit
 	s.pruningManager.PausePruning()
 
-	eg := new(errgroup.Group)
-
-	// commit SC async
-	var cInfo *proof.CommitInfo
-	eg.Go(func() error {
-		if err := s.stateCommitment.WriteChangeset(cs); err != nil {
-			return fmt.Errorf("failed to write batch to SC store: %w", err)
-		}
-		var scErr error
-		cInfo, scErr = s.stateCommitment.Commit(cs.Version)
-		if scErr != nil {
-			return fmt.Errorf("failed to commit SC store: %w", scErr)
-		}
-		return nil
-	})
-
-	if err := eg.Wait(); err != nil {
-		return nil, err
+	st := time.Now()
+	if err := s.stateCommitment.WriteChangeset(cs); err != nil {
+		return nil, fmt.Errorf("failed to write batch to SC store: %w", err)
 	}
+	writeDur := time.Since(st)
+	st = time.Now()
+	cInfo, err := s.stateCommitment.Commit(cs.Version)
+	if err != nil {
+		return nil, fmt.Errorf("failed to commit SC store: %w", err)
+	}
+	s.logger.Warn(fmt.Sprintf("commit version %d write=%s commit=%s", cs.Version, writeDur, time.Since(st)))
 
 	if cInfo.Version != cs.Version {
 		return nil, fmt.Errorf("commit version mismatch: got %d, expected %d", cInfo.Version, cs.Version)