+++ tests/integration/v2/app.go
@@ -17,6 +17,7 @@ import (
 	corebranch "cosmossdk.io/core/branch"
 	"cosmossdk.io/core/comet"
 	corecontext "cosmossdk.io/core/context"
+	"cosmossdk.io/core/gas"
 	"cosmossdk.io/core/header"
 	"cosmossdk.io/core/server"
 	corestore "cosmossdk.io/core/store"
@@ -99,10 +100,12 @@ type StartupConfig struct {
 	RouterServiceBuilder runtime.RouterServiceBuilder
 	// HeaderService defines the custom header service to be used in the app.
 	HeaderService header.Service
+
+	GasService gas.Service
 }
 
-func DefaultStartUpConfig(t *testing.T) StartupConfig {
-	t.Helper()
+func DefaultStartUpConfig(tb testing.TB) StartupConfig {
+	tb.Helper()
 
 	priv := secp256k1.GenPrivKey()
 	ba := authtypes.NewBaseAccount(
@@ -117,8 +120,8 @@ func DefaultStartUpConfig(t *testing.T) StartupConfig {
 			sdk.NewCoin(sdk.DefaultBondDenom, sdkmath.NewInt(100000000000000)),
 		),
 	}
-	homedir := t.TempDir()
-	t.Logf("generated integration test app config; HomeDir=%s", homedir)
+	homedir := tb.TempDir()
+	tb.Logf("generated integration test app config; HomeDir=%s", homedir)
 	return StartupConfig{
 		ValidatorSet:    CreateRandomValidatorSet,
 		GenesisBehavior: Genesis_COMMIT,
@@ -129,6 +132,7 @@ func DefaultStartUpConfig(t *testing.T) StartupConfig {
 			stf.NewMsgRouterService, stf.NewQueryRouterService(),
 		),
 		HeaderService: services.NewGenesisHeaderService(stf.HeaderService{}),
+		GasService:    stf.NewGasMeterService(),
 	}
 }
 
@@ -193,9 +197,11 @@ func NewApp(
 				startupConfig.BranchService,
 				startupConfig.RouterServiceBuilder,
 				startupConfig.HeaderService,
+				startupConfig.GasService,
 			),
 			depinject.Invoke(
 				std.RegisterInterfaces,
+				std.RegisterLegacyAminoCodec,
 			),
 		),
 		append(extraOutputs, &appBuilder, &cdc, &txConfigOptions, &txConfig, &storeBuilder)...); err != nil {
@@ -336,7 +342,7 @@ func (a *App) Deliver(
 	require.NoError(t, err)
 	a.lastHeight++
 
-	// update block height if integration context is present
+	// update block height and block time if integration context is present
 	iCtx, ok := ctx.Value(contextKey).(*integrationContext)
 	if ok {
 		iCtx.header.Height = int64(a.lastHeight)
@@ -345,10 +351,10 @@ func (a *App) Deliver(
 }
 
 // StateLatestContext creates returns a new context from context.Background() with the latest state.
-func (a *App) StateLatestContext(t *testing.T) context.Context {
-	t.Helper()
+func (a *App) StateLatestContext(tb testing.TB) context.Context {
+	tb.Helper()
 	_, state, err := a.Store.StateLatest()
-	require.NoError(t, err)
+	require.NoError(tb, err)
 	writeableState := branch.DefaultNewWriterMap(state)
 	iCtx := &integrationContext{state: writeableState}
 	return context.WithValue(context.Background(), contextKey, iCtx)
@@ -445,6 +451,8 @@ func (a *App) SignCheckDeliver(
 // It required the context to have the integration context.
 // a new state is committed if the option WithAutomaticCommit is set in options.
 func (app *App) RunMsg(t *testing.T, ctx context.Context, handler handler, option ...Option) (resp transaction.Msg, err error) {
+	t.Helper()
+
 	// set options
 	cfg := &RunMsgConfig{}
 	for _, opt := range option {