+++ tests/integration/v2/services.go
@@ -1,10 +1,13 @@
 package integration
 
 import (
+	"bytes"
 	"context"
+	"encoding/json"
 	"errors"
 	"fmt"
 
+	"github.com/cosmos/gogoproto/jsonpb"
 	gogoproto "github.com/cosmos/gogoproto/proto"
 
 	"cosmossdk.io/core/branch"
@@ -17,6 +20,8 @@ import (
 	"cosmossdk.io/core/server"
 	corestore "cosmossdk.io/core/store"
 	"cosmossdk.io/core/transaction"
+	"cosmossdk.io/server/v2/stf"
+	stfbranch "cosmossdk.io/server/v2/stf/branch"
 	stfgas "cosmossdk.io/server/v2/stf/gas"
 )
 
@@ -72,6 +77,7 @@ type integrationContext struct {
 	state    corestore.WriterMap
 	gasMeter gas.Meter
 	header   header.Info
+	events   []event.Event
 }
 
 func SetHeaderInfo(ctx context.Context, h header.Info) context.Context {
@@ -95,6 +101,45 @@ func SetCometInfo(ctx context.Context, c comet.Info) context.Context {
 	return context.WithValue(ctx, corecontext.CometInfoKey, c)
 }
 
+func EventsFromContext(ctx context.Context) []event.Event {
+	iCtx, ok := ctx.Value(contextKey).(*integrationContext)
+	if !ok {
+		return nil
+	}
+	return iCtx.events
+}
+
+func GetAttributes(e []event.Event, key string) ([]event.Attribute, bool) {
+	attrs := make([]event.Attribute, 0)
+	for _, event := range e {
+		attributes, err := event.Attributes()
+		if err != nil {
+			return nil, false
+		}
+		for _, attr := range attributes {
+			if attr.Key == key {
+				attrs = append(attrs, attr)
+			}
+		}
+	}
+
+	return attrs, len(attrs) > 0
+}
+
+func GetAttribute(e event.Event, key string) (event.Attribute, bool) {
+	attributes, err := e.Attributes()
+	if err != nil {
+		return event.Attribute{}, false
+	}
+	for _, attr := range attributes {
+		if attr.Key == key {
+			return attr, true
+		}
+	}
+
+	return event.Attribute{}, false
+}
+
 func GasMeterFromContext(ctx context.Context) gas.Meter {
 	iCtx, ok := ctx.Value(contextKey).(*integrationContext)
 	if !ok {
@@ -109,8 +154,17 @@ func GasMeterFactory(ctx context.Context) func() gas.Meter {
 	}
 }
 
+func SetGasMeter(ctx context.Context, meter gas.Meter) context.Context {
+	iCtx, ok := ctx.Value(contextKey).(*integrationContext)
+	if !ok {
+		return ctx
+	}
+	iCtx.gasMeter = meter
+	return context.WithValue(ctx, contextKey, iCtx)
+}
+
 func (s storeService) OpenKVStore(ctx context.Context) corestore.KVStore {
-	const gasLimit = 100_000
+	const gasLimit = 1_000_000
 	iCtx, ok := ctx.Value(contextKey).(*integrationContext)
 	if !ok {
 		return s.executionService.OpenKVStore(ctx)
@@ -133,19 +187,59 @@ var (
 type eventService struct{}
 
 // EventManager implements event.Service.
-func (e *eventService) EventManager(context.Context) event.Manager {
-	return &eventManager{}
+func (e *eventService) EventManager(ctx context.Context) event.Manager {
+	iCtx, ok := ctx.Value(contextKey).(*integrationContext)
+	if !ok {
+		panic("context is not an integration context")
+	}
+
+	return &eventManager{ctx: iCtx}
 }
 
-type eventManager struct{}
+type eventManager struct {
+	ctx *integrationContext
+}
 
 // Emit implements event.Manager.
-func (e *eventManager) Emit(event transaction.Msg) error {
+func (e *eventManager) Emit(tev transaction.Msg) error {
+	ev := event.Event{
+		Type: gogoproto.MessageName(tev),
+		Attributes: func() ([]event.Attribute, error) {
+			outerEvent, err := stf.TypedEventToEvent(tev)
+			if err != nil {
+				return nil, err
+			}
+
+			return outerEvent.Attributes()
+		},
+		Data: func() (json.RawMessage, error) {
+			buf := new(bytes.Buffer)
+			jm := &jsonpb.Marshaler{OrigName: true, EmitDefaults: true, AnyResolver: nil}
+			if err := jm.Marshal(buf, tev); err != nil {
+				return nil, err
+			}
+
+			return buf.Bytes(), nil
+		},
+	}
+
+	e.ctx.events = append(e.ctx.events, ev)
 	return nil
 }
 
 // EmitKV implements event.Manager.
 func (e *eventManager) EmitKV(eventType string, attrs ...event.Attribute) error {
+	ev := event.Event{
+		Type: eventType,
+		Attributes: func() ([]event.Attribute, error) {
+			return attrs, nil
+		},
+		Data: func() (json.RawMessage, error) {
+			return json.Marshal(attrs)
+		},
+	}
+
+	e.ctx.events = append(e.ctx.events, ev)
 	return nil
 }
 
@@ -173,13 +267,53 @@ func (bs *BranchService) ExecuteWithGasLimit(
 		return 0, errors.New("context is not an integration context")
 	}
 
+	originalGasMeter := iCtx.gasMeter
+
+	iCtx.gasMeter = stfgas.DefaultGasMeter(gasLimit)
+
 	// execute branched, with predefined gas limit.
-	err = f(ctx)
+	err = bs.execute(ctx, iCtx, f)
+
 	// restore original context
 	gasUsed = iCtx.gasMeter.Limit() - iCtx.gasMeter.Remaining()
-	_ = iCtx.gasMeter.Consume(gasUsed, "execute-with-gas-limit")
+	_ = originalGasMeter.Consume(gasUsed, "execute-with-gas-limit")
+	iCtx.gasMeter = stfgas.DefaultGasMeter(originalGasMeter.Remaining())
 
 	return gasUsed, err
+}
+
+func (bs BranchService) execute(ctx context.Context, ictx *integrationContext, f func(ctx context.Context) error) error {
+	branchedState := stfbranch.DefaultNewWriterMap(ictx.state)
+	meteredBranchedState := stfgas.DefaultWrapWithGasMeter(ictx.gasMeter, branchedState)
+
+	branchedCtx := &integrationContext{
+		state:    meteredBranchedState,
+		gasMeter: ictx.gasMeter,
+		header:   ictx.header,
+		events:   ictx.events,
+	}
+
+	newCtx := context.WithValue(ctx, contextKey, branchedCtx)
+
+	err := f(newCtx)
+	if err != nil {
+		return err
+	}
+
+	err = applyStateChanges(ictx.state, branchedCtx.state)
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func applyStateChanges(dst, src corestore.WriterMap) error {
+	changes, err := src.GetStateChanges()
+	if err != nil {
+		return err
+	}
+	return dst.ApplyStateChanges(changes)
 }
 
 // msgTypeURL returns the TypeURL of a proto message.
@@ -231,4 +365,16 @@ func (h *HeaderService) HeaderInfo(ctx context.Context) header.Info {
 		return header.Info{}
 	}
 	return iCtx.header
+}
+
+var _ gas.Service = &GasService{}
+
+type GasService struct{}
+
+func (g *GasService) GasMeter(ctx context.Context) gas.Meter {
+	return GasMeterFromContext(ctx)
+}
+
+func (g *GasService) GasConfig(ctx context.Context) gas.GasConfig {
+	return gas.GasConfig{}
 }